kyotei_predictor/requirements.txt
Flask==3.1.1
pandas==2.2.3
numpy
requests==2.32.3
python-dateutil==2.9.0.post0
metaboatrace.scrapers==3.3.1  # 競艇データ取得ライブラリ（必須）
Flask-Caching
psutil

# --- RL / Reinforcement Learning ---
gymnasium==0.29.1
stable-baselines3==2.2.1
torch==2.7.1
optuna==3.4.0
mlflow==2.8.1

# --- Testing ---
pytest==7.4.3
pytest-cov==4.1.0
pytest-html==4.1.1
